fix: スキルツリー編集の座標とCP対象クラスを修正

+++ 10_plugin/AstralRecord/src/main/java/io/github/maaasu/astralRecord/feature/skilltree/model/SkillTreePosition.java
@@ -14,9 +14,9 @@ import java.io.File;
 public record SkillTreePosition(
         @NotNull String nodeId,
         @NotNull String worldName,
-        int x,
-        int y,
-        int z
+        double x,
+        double y,
+        double z
 ) {
     @NotNull
     public String locationKey() {
